+++ includes/admin/class-cocart-admin-action-links.php
@@ -64,7 +64,7 @@ if ( ! class_exists( 'CoCart_Admin_Action_Links' ) ) {
 		 *
 		 * @access  public
 		 * @since   2.0.0
-		 * @version 2.7.2
+		 * @version 3.0.3
 		 * @param   array  $metadata An array of the plugin's metadata.
 		 * @param   string $file     Path to the plugin file.
 		 * @param   array  $data     Plugin Information
@@ -75,22 +75,27 @@ if ( ! class_exists( 'CoCart_Admin_Action_Links' ) ) {
 				/* translators: %s: URL to author */
 				$metadata[1] = sprintf( __( 'Developed By %s', 'cart-rest-api-for-woocommerce' ), '<a href="' . $data['AuthorURI'] . '" aria-label="' . esc_attr__( 'View the developers site', 'cart-rest-api-for-woocommerce' ) . '">' . $data['Author'] . '</a>' );
 
-				$campaign_args = array(
-					'utm_medium'   => 'cocart-lite',
-					'utm_source'   => 'plugins-page',
-					'utm_campaign' => 'plugins-row',
-					'utm_content'  => 'go-pro',
-				);
+				if ( ! CoCart_Helpers::is_cocart_pro_activated() ) {
+					$campaign_args = CoCart_Helpers::cocart_campaign( array(
+						'utm_content' => 'go-pro',
+					) );
+				} else {
+					$campaign_args = CoCart_Helpers::cocart_campaign( array(
+						'utm_content' => 'has-pro',
+					) );
+				}
+
+				$campaign_args['utm_campaign'] = 'plugins-row';
 
 				$row_meta = array(
 					/* translators: %s: CoCart */
-					'docs'      => '<a href="' . esc_url( COCART_DOCUMENTATION_URL ) . '" aria-label="' . sprintf( esc_attr__( 'View %s documentation', 'cart-rest-api-for-woocommerce' ), 'CoCart' ) . '" target="_blank">' . esc_attr__( 'Documentation', 'cart-rest-api-for-woocommerce' ) . '</a>',
-					'translate' => '<a href="' . esc_url( COCART_TRANSLATION_URL ) . '" aria-label="' . sprintf( esc_attr__( 'Translate %s', 'cart-rest-api-for-woocommerce' ), 'CoCart' ) . '" target="_blank">' . esc_attr__( 'Translate', 'cart-rest-api-for-woocommerce' ) . '</a>',
-					'review'    => '<a href="' . esc_url( COCART_REVIEW_URL ) . '" aria-label="' . sprintf( esc_attr__( 'Review %s on WordPress.org', 'cart-rest-api-for-woocommerce' ), 'CoCart' ) . '" target="_blank">' . esc_attr__( 'Leave a Review', 'cart-rest-api-for-woocommerce' ) . '</a>',
+					'docs'      => '<a href="' . CoCart_Helpers::build_shortlink( add_query_arg( $campaign_args, COCART_DOCUMENTATION_URL ) ) . '" aria-label="' . sprintf( esc_attr__( 'View %s documentation', 'cart-rest-api-for-woocommerce' ), 'CoCart' ) . '" target="_blank">' . esc_attr__( 'Documentation', 'cart-rest-api-for-woocommerce' ) . '</a>',
+					'translate' => '<a href="' . CoCart_Helpers::build_shortlink( add_query_arg( $campaign_args, COCART_TRANSLATION_URL ) ) . '" aria-label="' . sprintf( esc_attr__( 'Translate %s', 'cart-rest-api-for-woocommerce' ), 'CoCart' ) . '" target="_blank">' . esc_attr__( 'Translate', 'cart-rest-api-for-woocommerce' ) . '</a>',
+					'review'    => '<a href="' . CoCart_Helpers::build_shortlink( add_query_arg( $campaign_args, COCART_REVIEW_URL ) ) . '" aria-label="' . sprintf( esc_attr__( 'Review %s on WordPress.org', 'cart-rest-api-for-woocommerce' ), 'CoCart' ) . '" target="_blank">' . esc_attr__( 'Leave a Review', 'cart-rest-api-for-woocommerce' ) . '</a>',
 				);
 
-				// Only show donate option if CoCart Pro is not installed.
-				if ( ! CoCart_Helpers::is_cocart_pro_installed() ) {
+				// Only show donate option if CoCart Pro is not activated.
+				if ( ! CoCart_Helpers::is_cocart_pro_activated() ) {
 					$donate = array(
 						/* translators: %s: CoCart */
 						'donate'   => '<a href="' . esc_url( 'https://www.buymeacoffee.com/sebastien' ) . '" aria-label="' . sprintf( esc_attr__( 'Make a donation for %s', 'cart-rest-api-for-woocommerce' ), 'CoCart' ) . '" target="_blank" style="color: #399141; font-weight: 600;">' . esc_attr__( 'Donate', 'cart-rest-api-for-woocommerce' ) . '</a>',
@@ -100,8 +105,8 @@ if ( ! class_exists( 'CoCart_Admin_Action_Links' ) ) {
 					$row_meta = array_merge( $donate, $row_meta );
 				}
 
-				// Only show upgrade option if CoCart Pro is not installed.
-				if ( ! CoCart_Helpers::is_cocart_pro_installed() ) {
+				// Only show upgrade option if CoCart Pro is not activated.
+				if ( ! CoCart_Helpers::is_cocart_pro_activated() ) {
 					$store_url = CoCart_Helpers::build_shortlink( add_query_arg( $campaign_args, COCART_STORE_URL . 'pro/' ) );
 
 					/* translators: %s: CoCart Pro */