+++ includes/api/class-cocart-remove-item-controller.php
@@ -189,7 +189,9 @@ class CoCart_Remove_Item_v2_Controller extends CoCart_Item_Controller {
 	 * @return array $params
 	 */
 	public function get_collection_params() {
-		$params = array(
+		$controller = new CoCart_Cart_V2_Controller();
+
+		$params = array_merge( $controller->get_collection_params(), array(
 			'item_key'      => array(
 				'description'       => __( 'Unique identifier for the item in the cart.', 'cart-rest-api-for-woocommerce' ),
 				'type'              => 'string',
@@ -202,7 +204,7 @@ class CoCart_Remove_Item_v2_Controller extends CoCart_Item_Controller {
 				'type'              => 'boolean',
 				'validate_callback' => 'rest_validate_request_arg',
 			),
-		);
+		) );
 
 		return $params;
 	} // END get_collection_params()