+++ includes/class-cocart-cart-validation.php
@@ -6,6 +6,7 @@
  * @category Classes
  * @package  CoCart\Classes
  * @since    3.0.0
+ * @version  3.0.4
  * @license  GPL-2.0+
  */
 
@@ -31,22 +32,28 @@ if ( ! class_exists( 'CoCart_Cart_Validation' ) ) {
 		/**
 		 * Looks through the cart to check each item is in stock. If not, add error notice.
 		 *
-		 * @access public
-		 * @param  array  $cart_contents Cart contents before cart changes.
-		 * @param  WC_Cart               Cart object.
-		 * @return array  $cart_contents Cart contents after cart changes.
+		 * @access  public
+		 * @since   3.0.0
+		 * @version 3.0.4
+		 * @param   array  $cart_contents Cart contents before cart changes.
+		 * @param   WC_Cart               Cart object.
+		 * @return  array  $cart_contents Cart contents after cart changes.
 		 */
 		public function check_cart_item_stock( $cart_contents, $cart ) {
-			$product_qty_in_cart      = $cart->get_cart_item_quantities();
+			$qty_in_cart              = $cart->get_cart_item_quantities();
 			$current_session_order_id = isset( WC()->session->order_awaiting_payment ) ? absint( WC()->session->order_awaiting_payment ) : 0;
 
 			foreach ( $cart_contents as $item_key => $values ) {
 				$product = $values['data'];
 
+				$item_has_error = false;
+
 				// Check stock based on stock-status.
 				if ( ! $product->is_in_stock() ) {
 					/* translators: %s: product name */
-					wc_add_notice( sprintf( __( 'Sorry, "%s" is not in stock and cannot be purchased.', 'cart-rest-api-for-woocommerce' ), $product->get_name() ), 'error' );
+					wc_add_notice( sprintf( __( 'Sorry, "%s" is not in stock. Please edit your cart and try again. We apologize for any inconvenience caused.', 'cart-rest-api-for-woocommerce' ), $product->get_name() ), 'error' );
+
+					$item_has_error = true;
 				}
 
 				// We only need to check products managing stock, with a limited stock qty.
@@ -56,7 +63,7 @@ if ( ! class_exists( 'CoCart_Cart_Validation' ) ) {
 
 				// Check stock based on all items in the cart and consider any held stock within pending orders.
 				$held_stock     = wc_get_held_stock_quantity( $product, $current_session_order_id );
-				$required_stock = $product_qty_in_cart[ $product->get_stock_managed_by_id() ];
+				$required_stock = $qty_in_cart[ $product->get_stock_managed_by_id() ];
 
 				/**
 				 * Allows filter if product have enough stock to get added to the cart.
@@ -66,8 +73,10 @@ if ( ! class_exists( 'CoCart_Cart_Validation' ) ) {
 				 * @param array      $values    Cart item values.
 				 */
 				if ( apply_filters( 'cocart_cart_item_required_stock_is_not_enough', $product->get_stock_quantity() < ( $held_stock + $required_stock ), $product, $values ) ) {
-					/* translators: 1: product name 2: quantity in stock */
-					wc_add_notice( sprintf( __( 'Sorry, we do not have enough "%1$s" in stock to fulfill your order (%2$s available). We apologize for any inconvenience caused.', 'cart-rest-api-for-woocommerce' ), $product->get_name(), wc_format_stock_quantity_for_display( $product->get_stock_quantity() - $held_stock, $product ) ), 'error' );
+					if ( ! $item_has_error ) {
+						/* translators: 1: product name 2: quantity in stock */
+						wc_add_notice( sprintf( __( 'Sorry, we do not have enough "%1$s" in stock to fulfill your order (%2$s available). We apologize for any inconvenience caused.', 'cart-rest-api-for-woocommerce' ), $product->get_name(), wc_format_stock_quantity_for_display( $product->get_stock_quantity() - $held_stock, $product ) ), 'error' );
+					}
 				}
 			}