+++ includes/class-cocart.php
@@ -26,7 +26,7 @@ final class CoCart {
 	 * @access public
 	 * @static
 	 */
-	public static $version = '3.0.0';
+	public static $version = '3.0.4';
 
 	/**
 	 * CoCart Database Schema version.